Update status.rst
Hyperlink Formatting: In the sections "What is the fee for previous blocks?" and "What can I do if the fee is too high?", the hyperlinks are not formatted correctly for plain text. It's better to provide the full URL next to the text it's meant to link. For example:

For the fee for previous blocks, use getFeeAmount(): [https://github.com/ethereum/btcrelay/tree/master#getfeeamountblockhash]
To change the fee recipient, use changeFeeRecipient(): [https://github.com/ethereum/btcrelay/tree/master#changefeerecipientblockhash-fee-recipient]